Calculates OpList based Size pre MAX
This is as small as i could make the code work. Basically i don't use the Image Size (source) directly, i request from the Operation List a "size so far". And from there on Max/Scaling works on that new Size (if there is any. E.g FULL will give me same as the Source Size/no cropping). This also requires that the scaled Size is based on that, since (another side effect of the previous code) Crop Operations in excess (e.g requesting more pixels than present, which is IIIF valid) would end with a scaling that would not respect the original cropping)