Ok. We can't just use OpList Method because it can throw an internal exception
If the arguments passed for crop are not the right ones.
So we have to be more selective (so much for short/clean code), just pick the Crop operation (if any), validate (so we can throw a catchable exception later on) and then request the resulting size only for that one.
Don't love reusing requestedSize (bc of read-ability) but better than creating yet another Objects